fix(configitems): move obproxy config items to obproxy models
