+++ internal/dashboard/handler/alarm_handler.go
@@ -14,6 +14,8 @@ package handler
 
 import (
 	"github.com/gin-gonic/gin"
+	"github.com/mcuadros/go-defaults"
+
 	"github.com/oceanbase/ob-operator/internal/dashboard/business/alarm"
 	"github.com/oceanbase/ob-operator/internal/dashboard/model/alarm/alert"
 	"github.com/oceanbase/ob-operator/internal/dashboard/model/alarm/receiver"
@@ -21,8 +23,6 @@ import (
 	"github.com/oceanbase/ob-operator/internal/dashboard/model/alarm/rule"
 	"github.com/oceanbase/ob-operator/internal/dashboard/model/alarm/silence"
 	httpErr "github.com/oceanbase/ob-operator/pkg/errors"
-
-	"github.com/mcuadros/go-defaults"
 )
 
 // @ID ListAlerts
