+++ internal/dashboard/handler/obproxy_handler.go
@@ -17,7 +17,6 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	biz "github.com/oceanbase/ob-operator/internal/dashboard/business/obproxy"
-	"github.com/oceanbase/ob-operator/internal/dashboard/model/common"
 	"github.com/oceanbase/ob-operator/internal/dashboard/model/obproxy"
 	"github.com/oceanbase/ob-operator/internal/dashboard/model/param"
 	httpErr "github.com/oceanbase/ob-operator/pkg/errors"
@@ -145,12 +144,12 @@ func DeleteOBProxy(c *gin.Context) (*obproxy.OBProxy, error) {
 // @Produce application/json
 // @Param namespace path string true "namespace of obproxy deployment"
 // @Param name path string true "name of obproxy deployment"
-// @Success 200 object response.APIResponse{data=[]common.ConfigItem}
+// @Success 200 object response.APIResponse{data=[]obproxy.ConfigItem}
 // @Failure 400 object response.APIResponse
 // @Failure 401 object response.APIResponse
 // @Failure 500 object response.APIResponse
 // @Router /api/v1/obproxies/{namespace}/{name}/parameters [GET]
-func ListOBProxyParameters(c *gin.Context) ([]common.ConfigItem, error) {
+func ListOBProxyParameters(c *gin.Context) ([]obproxy.ConfigItem, error) {
 	nn := &param.NamespacedName{}
 	err := c.BindUri(nn)
 	if err != nil {
